TestFreezeDetector now uses TestFX API to run unit tests
