+++ freeze-detector/src/main/java/one/jpro/platform/freezedetector/FreezeDetector.java
@@ -5,7 +5,6 @@ import javafx.application.Platform;
 
 import java.time.Duration;
 import java.util.function.BiConsumer;
-import java.util.function.Consumer;
 
 public class FreezeDetector {
